+++ estadisticas.c
@@ -12,7 +12,7 @@
 #define MEM_SIZE 15     // Tamaño de la memoria compartida en bytes
 
 sem_t semaforo;         // instancia de semaforo
-int *ptr_entero;        // puntero para la memoria compartida
+double *ptr_entero;        // puntero para la memoria compartida
 
 int main() {
 
@@ -41,7 +41,7 @@ int main() {
         exit(1);
     }
 
-    ptr_entero = (int *)memoria;    // ver si debo poner el asterisco antes de ptr
+    ptr_entero = (double *)memoria;    // ver si debo poner el asterisco antes de ptr
     printf("Esperando que los procesos terminen.... \n\n");
 
     while (1)
@@ -54,9 +54,9 @@ int main() {
             printf("--------> Estadísticas de los procesos --------> \n\n");
             printf("Tiempo bloqueado del Cliente: %0.10f segundos.\n", ptr_entero[2]);              //
             printf("Tiempo bloquedado del Reconstructor: %0.10f segundos.\n", ptr_entero[3]);       // server
-            printf("Número de caracteres transferidos: %d caracteres.\n", ptr_entero[0]);           //
-            printf("Número de caracteres en el buffer: %d caracteres.\n", ptr_entero[4]);           // server
-            printf("Espacio total de memoria utilizado: %d bytes.\n", ptr_entero[5]);               //
+            printf("Número de caracteres transferidos: %f caracteres.\n", ptr_entero[0]);           //
+            printf("Número de caracteres en el buffer: %f caracteres.\n", ptr_entero[4]);           // server
+            printf("Espacio total de memoria utilizado: %f bytes.\n", ptr_entero[5]);               //
             printf("Tiempo del Cliente en modo Usuario: %0.10f segundos.\n", ptr_entero[6]);
             printf("Tiempo del Cliente en modo Kernel: %0.10f segundos.\n", ptr_entero[7]);         //
             printf("Tiempo del Reconstructor en modo Usuario: %0.10f segundos.\n", ptr_entero[8]);
